pubsub_football/requirements.txt
google-cloud-pubsub
requests
flask